refactor: move parseSegments to render-template.ts, add aria-label (#11)
